Refactor conductor code structure
