+++ conductor/src/types/ConductorLog.ts
@@ -1,0 +1,7 @@
+type ConductorLog = {
+  createdTime?: number
+  log?: string
+  taskId?: string
+}
+
+export default ConductorLog
